requirements.txt
absl-py==1.0.0
appdirs==1.4.4
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cityscapesScripts==2.2.0
coloredlogs==15.0.1
cycler==0.11.0
fonttools==4.29.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.43.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.0
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
pyquaternion==0.9.9
python-dateutil==2.8.2
pytorch-ignite==0.4.9
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
typing==3.7.4.3
typing_extensions==4.1.0
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.7.0
